+++ lib/ui/home/HomeInboxWidget.dart
@@ -21,6 +21,8 @@ import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
+enum HomeInboxContent { all, unread }
+
 class HomeInboxWidget extends StatefulWidget {
 
   final HomeInboxContent content;
@@ -62,6 +64,10 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
   Key _pageViewKey = UniqueKey();
   Map<String, GlobalKey> _contentKeys = <String, GlobalKey>{};
   final double _pageSpacing = 16;
+
+  static const String localScheme = 'local';
+  static const String allNotificationsHost = 'all_notifications';
+  static const String localUrlMacro = '{{local_url}}';
 
   @override
   void initState() {
@@ -211,9 +217,19 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     else if (_loadingMessages) {
       return HomeProgressWidget();
     }
+    else if (_messages == null) {
+      return HomeMessageCard(message: Localization().getStringEx("widget.home.inbox.text.failed.description", "Failed to load notifications."),);
+    }
     else if (CollectionUtils.isEmpty(_messages)) {
-      return HomeMessageCard(
-        message: Localization().getStringEx("widget.home.inbox.text.empty.description", "Notifications are available right now."),);
+      if (widget.content == HomeInboxContent.unread) {
+        String message = Localization().getStringEx("widget.home.inbox.text.empty.unread.description", "You have no unread notifications. <a href='$localUrlMacro'><b>View all notifications</b></а>.")
+          .replaceAll(localUrlMacro, '$localScheme://$allNotificationsHost');
+          return HomeMessageHtmlCard(message: message, onTapLink: _onTapMessageLink,);
+      }
+      else {
+        return HomeMessageCard(
+          message: Localization().getStringEx("widget.home.inbox.text.empty.all.description", "You have not any notifications yet."),);
+      }
     }
     else {
       return _buildMessagesContent();
@@ -285,6 +301,15 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     return minContentHeight ?? 0;
   }
 
+  void _onTapMessageLink(String? url) {
+    Uri? uri = (url != null) ? Uri.tryParse(url) : null;
+    if (uri?.scheme == localScheme) {
+      if (uri?.host.toLowerCase() == allNotificationsHost.toLowerCase()) {
+        SettingsNotificationsContentPanel.present(context, content: SettingsNotificationsContent.all);
+      }
+    }
+  }
+
   void _onTapMessage(InboxMessage message) {
     Analytics().logSelect(target: message.subject);
     SettingsNotificationsContentPanel.launchMessageDetail(message);
@@ -295,5 +320,3 @@ class _HomeInboxWidgetState extends State<HomeInboxWidget> implements Notificati
     SettingsNotificationsContentPanel.present(context, content: (_unread == true) ? SettingsNotificationsContent.unread : SettingsNotificationsContent.all);
   }
 }
-
-enum HomeInboxContent { all, unread }